Swapping control screen for control sheet
