+++ lib/helper_widgets/header.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+class Header extends StatelessWidget {
+  const Header(this.title, {this.subtitle, super.key});
+
+  final String title;
+  final String? subtitle;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 16.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(title,
+              style: Theme.of(context)
+                  .textTheme
+                  .headlineSmall!
+                  .copyWith(color: Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.7))),
+          if (subtitle != null) const SizedBox(height: 2),
+          if (subtitle != null)
+            Text(subtitle!,
+                style: Theme.of(context)
+                    .textTheme
+                    .titleMedium!
+                    .copyWith(color: Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.7))),
+        ],
+      ),
+    );
+  }
+}
