+++ lib/stats/settings_section.dart
@@ -12,10 +12,10 @@ import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import '../control_screen.dart';
 import '../domain/theme_helper.dart';
 import '../domain/scooter_keyless_distance.dart';
 import '../scooter_service.dart';
+import '../helper_widgets/header.dart';
 
 class SettingsSection extends StatefulWidget {
   const SettingsSection({super.key});
